move ExpressionUtils to package-private
